+++ src/Modules/BlizzardAPI.lua
@@ -259,7 +259,7 @@ end
 addon.GetItemCategory = GetItemCategory
 
 local function GetTransmogLocation(itemLinkOrID)
-	return TransmogUtil.GetTransmogLocation(GetItemSlot(itemLinkOrID), Enum.TransmogType.Appearance, Enum.TransmogModification.Main)
+	return TransmogUtil.GetTransmogLocation(GetItemSlot(itemLinkOrID), Enum.TransmogType.Appearance, false)
 end
 addon.GetTransmogLocation = GetTransmogLocation